requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
strawberry-graphql[fastapi]>=0.232.0
motor>=3.3.0
pydantic-settings>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
qdrant-client